Extract MsgHandler to handler.go
